requirements.txt
pandas
numpy
scikit-learn
lightgbm
mlflow
fastapi
uvicorn
pyyaml
scipy
pydantic
requests
implicit